requirements.txt
requests
fastapi
psycopg2-binary
jinja2
python-multipart
